Fixed Alignment Issue & Added Hover Effects to the cards
