--- src/components/card_aboutPage/ClubCard.css
@@ -1,52 +1,0 @@
-.club_card {
-  box-shadow: 0px 12px 37px rgba(26, 88, 155, 0.25);
-  border-radius: 16px 16px 16px 16px;
-}
-
-/* .club_card {
-  position: relative;
-  width: 300px;
-  height: 200px;
-  border-radius: 10px;
-  box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.5);
-  transition: 0.3s;
-  padding: 30px 50px;
-  background: #fff;
-  cursor: pointer;
-  margin-right: 30px;
-}
-
-.club_card:hover {
-  height: 320px;
-}
-
-.club_imgbox {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  transform: translateY(-80px);
-  z-index: 99;
-}
-
-.club_img {
-  width: 100%;
-  border-radius: 10px;
-  box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.5);
-}
-
-.club_content {
-  padding: 10px 20px;
-  text-align: center;
-  transform: translateY(-450px);
-  opacity: 0;
-  transition: 0.3s;
-}
-
-.club_card:hover > .club_content {
-  opacity: 1;
-  transform: translateY(-180px);
-}
-
-.club_content h2 {
-  color: #7f9ead;
-} */
