+++ src/pages/AboutPage/AboutPage.css
@@ -1,0 +1,7 @@
+.clubBox{
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+   margin: 20px 20px;
+    justify-content: space-evenly;
+}
